Misc. content updates, add more relevant content, fix typos, add rel='nofollow' for 3P software/service links.
nofollow is recommended for links we are not sure of (user submitted), we can also use rel='ugc' but some of these links won't entirely be user submitted.
We don't necessarily want to lose credibility in Google's search algorithms to point to  3P sites that in future 404 so rel='nofollow' is a safe bet to add here. I also checked Capterra is adding rel='nofollow'. We can later split this to sponsored/ugc tags but I don't see the need for that yet.