Update price filter field name
Why: Because we are renaming pay_now_price to stripe_primary_price on the backend side: https://github.com/taggedweb/taggedweb/pull/612
This rename is being made because for metered solutions there is no concept of "pay now". Metered solutions are billed later.
stripe_primary_price is a unified price applicable for both metered/pre-paid solutions. Note that this PR only changes the filter, we still have to change other references of pay_now_price (that we can do in a separate PR). It's important to change the filter first because that field is already updated on the backend side (Other pay_now_price_* fields defined in serializer on the backend are deprecated but not removed yet so they can be removed in a separate PR)